Bug fix: Now the file keeper is able to handle file deletions.
If a watched file is deleted, we must delete its hard link as well and
commit the change set. However as simple as it sounds, I had to change
the way that the file_keeper.c fetch the file name. It's impossible to
use g_file_quer_info() to fetch the file's name, since the file
does not exist anymore. I fixed this problem by adding a function
that returns the file name based on its path.