Bug fix: Handling binary files editing.
When a binary file is edited and saved the GFileMonitor reports that
the operation type is G_FILE_MONITOR_EVENT_DELETED, makig us belive
that the file was deleted. However the file may still be there, so
to properly handle binary file changes we must check if the file
was really deleted, because if it was not we should register another
monitor to it, recreate the hard link and commit the changes!