Bug fix: Stop rescheduling the timer, because in some scenarios the timeout may never happen!
Imagine a case where constant file changes are being made in the
file keeper's dir, with the old logic we were always rescheduling the
timer, delaying its execution. If we create an program that won't
stop to write in the file kepeer's dir the timeout would never expire
and the changes would never be saved. From now on only create the timer
if the previous one has expired!