Make it faster !
Avoid list usage for search, this list can be pretty big.

Also i removed that comment /* Don't even start */ and the code bellow it,
because the list (now the hash) can be empty at the program startup.
This case can be consired that all files were deleted from the file
keeper's root dir