Explicitly cancel the monitor watch using g_file_monitor_cancel().
If the monitor is deleted without calling this fuction the program may
segfault in a very specific case. When the user is editing the same binary
file constantly, a lot of "delete" events will happen and we will delete the monitor and create a new one. After some time we will stop receiving file change
notification and the program will segfault at glib level.

The curious thing is, I supposed that when a GFileMonitor is deleted
it should cancel its monitoring file by itself. Strange....