Bug fix: Chaging the rule to add File_Changed_Info to the _changed_files list.
There was a problem when the user edited a file and right before he
deleted the file. We would not insert a new changed file info because
the list _changed_files already has an entry with that hash, however
this entry is not marked as deleted and the file keeper will try to
commit files changes and it will fail, since the file does not exist anymore.
This patch addresses this kind of problem.