Adding a FileConn object.
The FileConn class is responsible for handling client connections.
These client connections are GUI applications that will display, the
user's files and will help the user to revert his/her files to older versions

The main function of the FileConn class is to notify the file keeper
daemon that a user is online and receive/send messages from/to the user.

For now the FileConn has a barebones implementation and we will need
to implement the user commands to list files and revert files, the daemon
should also notify the user that a file was added, updated
(there is a new commit for a monitored file) or deleted so the GUI can
display the new info.

 - TODO -

- Create a message protocol

- The following Commands should be implemented by the File Keeper daemon:

* From Daemon to User
* FILE_LIST -> Will tell the user which files we are monitoring.
* FILE_UPDATED -> Will tell the user that a given file has a new commit. (So the GUI can be updated)
* FILE_NEW -> A new file was added to the daemon monitoring. (So the GUI can be updated)
* FILE_DELETED -> A File was deleted (So the GUI can be updated)

* From User to File keeper daemon
* FILE_REVERT -> The user wants to revert a file to a given date