Implementing FILE_MESSAGE_REVERT_CONFIRM and FILE_MESSAGE_REVERT_ABORT.
When the user is reverting the file, he/she can preview its content before
making a decision. So in order to tell the daemon that a decisons has been
made we must use the command FILE_MESSAGE_REVERT_CONFIRM to reset the file
to the current commit or FILE_MESSAGE_REVERT_ABORT leave the file as it was.