Memory leak fix at file_keeper_revert_file().
As a user it is possible to keep asking the daemon to revert the file
to older versions.  The FileKeeper struct holds a git_commit object
that points to the last reverted commit, if the user revert the file again
we must free the previous commit, otherwise we will leak it.